Split search formatting modules
